vsclib: refactor allocator, add resource support
Bit difficult to split this up, so putting it all-in-one.

Switch the allocator code to internally use C++, in order to gain access
to templates, allowing us to use a different header structure.

Then, implement BRender-inspired resources. Resources are simply a block
of memory with that can have children. When a resource is freed, its children
are also recursively freed, simplifying ownership.

Each vsc_res_*() function takes a void* as the first argument - this can
either be NULL (to allocate a new hierarchy), or a previously-returned
value of vsc_res_allocate() (to attach the new block).

A VscAllocator instance for a given resource may be obtained via the
vsc_res_allocator() call. Passing this to an allocator-aware function
will cause any resources allocated with it to be attached to the owner
of the allocator, e.g:

> const VscAllocator a = vsc_res_allocator(res);
> return vsc_strdupa(s, &a);

Some helpers, e.g. vsc_res_strdup(), are provided for convenience.